requirements.txt
torch>=2.0.1
pesq==0.0.4
pystoi==0.3.3
scipy>=1.10.1
soundfile==0.12.1
torchaudio>=2.0.2
yamlargparse==1.31.1
torchinfo==1.8.0
tqdm==4.67.0
pyyaml
pysptk
pymcd
pyworld
fastdtw
museval
resampy
onnxruntime
gammatone
librosa==0.10.2.post1
opencv-python==4.10.0.84
mir_eval==0.7
numpy>=1.24.3
tensorboard==2.14.0
torch-complex>=0.4.4
rotary-embedding-torch==0.8.3
python_speech_features==0.6
typeguard==4.3.0
scenedetect
torchvision
huggingface-hub>=0.26.2
gdown==5.2.0
pydub
xls_r_sqa
wandb>=0.16.0
datasets>=2.14.0
pyarrow>=8.0.0
google-cloud-storage>=2.10.0
webrtcvad>=2.0.10
